# apidocs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "ChangePasswordRequest": {
        "properties": {
          "current_password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Current Password",
            "type": "string"
          },
          "new_password": {
            "maxLength": 72,
            "minLength": 6,
            "title": "New Password",
            "type": "string"
          },
          "new_password2": {
            "maxLength": 72,
            "minLength": 6,
            "title": "New Password2",
            "type": "string"
          }
        },
        "required": [
          "current_password",
          "new_password",
          "new_password2"
        ],
        "title": "ChangePasswordRequest",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "\u7edf\u4e00\u9519\u8bef\u54cd\u5e94\u683c\u5f0f\uff08Pinned Error Contract\uff09\u3002\n\n\u8be5\u7ed3\u6784\u6700\u521d\u7528\u4e8e v2\uff0c\u4f46\u5728\u201c\u5408\u5e76 v2 \u5230 v1\uff08\u4ec5\u4fdd\u7559 /api/v1\uff09\u201d\u540e\uff0c\n\u670d\u52a1\u7aef\u5bf9\u5916\u7edf\u4e00\u4f7f\u7528\u6b64\u9519\u8bef\u54cd\u5e94\u683c\u5f0f\uff0c\u4fbf\u4e8e\u5ba2\u6237\u7aef\u4e0e Web \u7aef\u505a\u4e00\u81f4\u7684\u9519\u8bef\u89e3\u6790\u3002",
        "properties": {
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Request Id"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 72,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
      "RegisterRequest": {
        "properties": {
          "password": {
            "maxLength": 72,
            "minLength": 6,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Flow Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login_api_v1_auth_login_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Clear the user session cookie.\n\n- Idempotent: OK even when already logged out.\n- If a valid cookie-session exists, enforce CSRF to prevent cross-site logout.\n- Bearer-token clients should not need CSRF for this endpoint.",
        "operationId": "logout_api_v1_auth_logout_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "register_api_v1_auth_register_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests",
            "headers": {
              "Retry-After": {
                "description": "Seconds to wait before retrying.",
                "schema": {
                  "type": "string"
                }
              },
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "summary": "Register",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "operationId": "get_me_api_v1_me_get",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Me",
        "tags": [
          "me"
        ]
      }
    },
    "/api/v1/me/password": {
      "post": {
        "operationId": "change_password_api_v1_me_password_post",
        "parameters": [
          {
            "description": "Optional client-provided request id; echoed back in responses.",
            "in": "header",
            "name": "X-Request-Id",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Validation Error",
            "headers": {
              "X-Request-Id": {
                "description": "Echoed or generated request id.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "me"
        ]
      }
    },
...
  }
}